Renamed Authorization to API. Will impliment api functionality
