+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 // import './scss/main.scss';
 import './scss/tailwind.css';
-import Authorization from './features/Authorization/Authorization';
+import Authorization from './features/api/api';
 import Menu from './components/navigation/Menu';
 /**
  * @license     MIT License
